+++ R/doPCA.R
@@ -22,6 +22,8 @@
 #'     the size of the labels will be scaled down in an attempt to keep the
 #'     labels inside the canvas. Typically set to half the width of the
 #'     plot device (in inches).
+#' @param colourBy Character scalar indicating the name of the column of
+#'     \code{colData(sce)} that will be used to colour the points.
 #'
 #' @export
 #' @author Charlotte Soneson
@@ -45,14 +47,14 @@
 #'     geom_text element_blank element_line scale_fill_gradient2 theme_bw
 #' @importFrom cowplot theme_cowplot plot_grid
 #' @importFrom BiocSingular ExactParam
-#' @importFrom SummarizedExperiment rowData
+#' @importFrom SummarizedExperiment rowData colData
 #' @importFrom SingleCellExperiment reducedDim
 #' @importFrom GGally ggpairs
 #' @importFrom utils head tail
 #'
 doPCA <- function(sce, assayName, ncomponents = 10, ntop = Inf,
                   plotpairs = list(c(1, 2)), maxNGroups = 10,
-                  maxTextWidthBarplot = NULL) {
+                  maxTextWidthBarplot = NULL, colourBy = "group") {
 
     ## ---------------------------------------------------------------------- ##
     ## Check arguments
@@ -65,6 +67,8 @@ doPCA <- function(sce, assayName, ncomponents = 10, ntop = Inf,
     .assertVector(x = plotpairs, type = "list")
     .assertScalar(x = maxNGroups, type = "numeric")
     .assertScalar(x = maxTextWidthBarplot, type = "numeric", allowNULL = TRUE)
+    .assertScalar(x = colourBy, type = "character", allowNULL = TRUE,
+                  validValues = colnames(SummarizedExperiment::colData(sce)))
     for (elm in plotpairs) {
         .assertVector(x = elm, type = "numeric", len = 2)
     }
@@ -115,14 +119,18 @@ doPCA <- function(sce, assayName, ncomponents = 10, ntop = Inf,
     for (pr in plotpairs) {
         pcadf <- data.frame(
             sampleLabel = colnames(sce),
-            SingleCellExperiment::reducedDim(sce, paste0("PCA_", assayName))[, pr],
-            group = sce$group)
+            SingleCellExperiment::reducedDim(sce, paste0("PCA_", assayName))[, pr])
+        if (!is.null(colourBy)) {
+            pcadf[[colourBy]] <- SummarizedExperiment::colData(sce)[[colourBy]]
+        }
         p <- ggplot2::ggplot(pcadf,
                              ggplot2::aes(x = .data[[paste0("PC", pr[1])]],
                                           y = .data[[paste0("PC", pr[2])]],
-                                          color = .data$group,
-                                          label = .data$sampleLabel)) +
-            ggplot2::geom_point(alpha = 0.5, size = 4) +
+                                          label = .data$sampleLabel))
+        if (!is.null(colourBy)) {
+            p <- p + ggplot2::aes(color = .data[[colourBy]])
+        }
+        p <- p + ggplot2::geom_point(alpha = 0.5, size = 4) +
             ggplot2::ggtitle(paste0("Assay: ", assayName)) +
             ggplot2::coord_fixed() +
             ggplot2::labs(
@@ -209,14 +217,18 @@ doPCA <- function(sce, assayName, ncomponents = 10, ntop = Inf,
         )
 
         ## Combine
-        pscat <- plist[[paste0("PC", pr[1], "_", pr[2])]] +
-            ggalt::geom_encircle(
-                ggplot2::aes(fill = .data$group, group = .data$group),
-                alpha = 0.5, show.legend = FALSE, na.rm = TRUE,
-                s_shape = 0.5, expand = 0.05, spread = 0.1)
-        if (length(unique(pcadf$group)) > maxNGroups) {
+        pscat <- plist[[paste0("PC", pr[1], "_", pr[2])]]
+        if (!is.null(colourBy)) {
             pscat <- pscat +
-                theme(legend.position = "none")
+                ggalt::geom_encircle(
+                    ggplot2::aes(fill = .data[[colourBy]],
+                                 group = .data[[colourBy]]),
+                    alpha = 0.5, show.legend = FALSE, na.rm = TRUE,
+                    s_shape = 0.5, expand = 0.05, spread = 0.1)
+            if (length(unique(pcadf[[colourBy]])) > maxNGroups) {
+                pscat <- pscat +
+                    theme(legend.position = "none")
+            }
         }
         plistcomb[[paste0("PC", pr[1], "_", pr[2])]] <-
             cowplot::plot_grid(
@@ -230,14 +242,25 @@ doPCA <- function(sce, assayName, ncomponents = 10, ntop = Inf,
     ## ---------------------------------------------------------------------- ##
     ## Pairs plot
     ## ---------------------------------------------------------------------- ##
-    pairsdf <- scuttle::makePerCellDF(sce, use.dimred = paste0("PCA_", assayName),
-                                      use.coldata = "group")
-    pcidx <- which(colnames(pairsdf) != "group")
-    colnames(pairsdf) <- sub(paste0("PCA_", assayName, "."), "PC",
-                             colnames(pairsdf))
-    ppairs <- GGally::ggpairs(pairsdf, columns = pcidx,
-                              ggplot2::aes(colour = .data$group),
-                              upper = "blank") +
+    if (!is.null(colourBy)) {
+        pairsdf <- scuttle::makePerCellDF(sce, use.dimred = paste0("PCA_", assayName),
+                                          use.coldata = colourBy)
+        pcidx <- which(colnames(pairsdf) != colourBy)
+        colnames(pairsdf) <- sub(paste0("PCA_", assayName, "."), "PC",
+                                 colnames(pairsdf))
+        ppairs <- GGally::ggpairs(pairsdf, columns = pcidx,
+                                  ggplot2::aes(colour = .data[[colourBy]]),
+                                  upper = "blank")
+    } else {
+        pairsdf <- scuttle::makePerCellDF(sce, use.dimred = paste0("PCA_", assayName),
+                                          use.coldata = FALSE)
+        pcidx <- seq_len(ncol(pairsdf))
+        colnames(pairsdf) <- sub(paste0("PCA_", assayName, "."), "PC",
+                                 colnames(pairsdf))
+        ppairs <- GGally::ggpairs(pairsdf, columns = pcidx,
+                                  upper = "blank")
+    }
+    ppairs <- ppairs +
         ggplot2::theme_bw() +
         ggplot2::ggtitle(paste0("Assay: ", assayName))